Streamlit frontend

+++ src/assets/producer.py
@@ -6,7 +6,7 @@ from assets.energy_asset import EnergyAsset
 class Producer(EnergyAsset, ABC):
     """Parent class for all power producers"""
 
-    def __init__(self, name: str, max_capacity: float, efficiency: float, asset_id: int = None):
+    def __init__(self, name: str, max_capacity: int, efficiency: float, asset_id: int = None):
         super().__init__(name = name, is_connected=True, asset_id=asset_id)
         self.max_capacity = max_capacity # in kW
         self.efficiency = efficiency
